move comparisons corrections to separate subdirectory
