+++ statannot/comparisons_corrections/ComparisonsCorrection.py
@@ -1,4 +1,4 @@
-from .comparisons_corrections import bonferroni, holm_bonferroni, benjamin_hochberg
+from statannot.comparisons_corrections.funcs import bonferroni, holm_bonferroni, benjamin_hochberg
 
 methods_names = {'bonferroni':        'Bonferroni',
                  'bonf':              'Bonferroni',
